pyproject.toml
[project]
name = "pptgen"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "NabeRoute", email = "kouya1278@gmail.com" }
]
dependencies = [
    "openai<=1.7.0",
    "streamlit<=1.29.0",
    "python-pptx<=0.6.23",
    "pandas<=1.3.5",
    "matplotlib<=3.4.3",
    "python-dotenv<=0.20.0",
    "networkx<=3.2",
    "setuptools<=68.0.0",
    "scipy<=1.11.1",
    "llmx>=0.0.21a0",
    "lida>=0.0.14",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pptgen"]
